feat: token 보유시 로그인 정보 획득 위치 변경
token 보유시 token으로 회원정보 획득 위치 최상단으로 이동
유저 프로필 페이지에 auth.guard 적용
