+++ src/app/auth.guard.ts
@@ -1,15 +1,19 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree, CanActivate, Router } from '@angular/router';
-import { Observable } from 'rxjs';
+import { CanActivate, Router } from '@angular/router';
+import { StorageService } from './core/services/storage.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AuthGuard implements CanActivate {
-  constructor(private router: Router) { }
+  constructor(
+    private router: Router
+    , private storageService: StorageService
+  ) { }
   canActivate() {
-    if (localStorage.getItem('my-token'))
+    if (this.storageService.getLocal("user")) {
       return true;
+    }
     else {
       this.router.navigate(['signin']);
       return false;
