+++ src/app/user/components/users.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { UserService } from 'src/app/core/services/user.service';
+import { CommonService } from 'src/app/core/services/common.service';
 
 @Component({
   selector: 'app-users',
@@ -13,7 +14,7 @@ import { UserService } from 'src/app/core/services/user.service';
             <div class="profile_share"></div>
             <div class="profile_pic"></div>
             <div class="profile_info">
-              <div class="profile_name"><strong>#NAME</strong></div>
+              <div class="profile_name"><strong>{{this.commonService.getUserDetail() ? this.commonService.getUserDetail()['username'] : ''}}</strong></div>
               <div class="profile_follow_state">
                 <div><a href="#"></a>팔로워0</div>
                 <div><a href="#"></a>팔로잉0</div>
@@ -232,7 +233,10 @@ import { UserService } from 'src/app/core/services/user.service';
   ]
 })
 export class UsersComponent implements OnInit {
-  constructor(private userService: UserService) {}
+  constructor(
+    private userService: UserService
+    , private commonService: CommonService
+  ) { }
 
-  ngOnInit() {}
+  ngOnInit() { }
 }
