+++ src/app/user/user-routing.module.ts
@@ -3,11 +3,12 @@ import { Routes, RouterModule } from '@angular/router';
 import { UsersComponent } from './components/users.component';
 import { OrderListComponent } from './components/order-list.component';
 import { UserModifyComponent } from './components/user-modify.component';
+import { AuthGuard } from '../auth.guard';
 
 const routes: Routes = [
-  { path: 'users/:id', component: UsersComponent },
-  { path: 'orderList/:id', component: OrderListComponent },
-  { path: 'users/:id/edit', component: UserModifyComponent },
+  { path: 'users/:id', component: UsersComponent, canActivate: [AuthGuard] },
+  { path: 'orderList/:id', component: OrderListComponent, canActivate: [AuthGuard] },
+  { path: 'users/:id/edit', component: UserModifyComponent, canActivate: [AuthGuard] },
 ];
 
 @NgModule({
